bugfix when checking invalid whl file
